Reorganize Nonlinearity Index Code
